refactor(works): add pre-publish concept
